requirements/development.txt
-r base.txt
django-tailwind==3.8.0
factory-boy==3.3.3
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
